chore: Add npm maintainers
This adds a maintainers field to the package.json. Ideally, someone else should be in this field, too. Please add another contributor as a maintainer for this npm package, such as @shamb0t, by typing: `npm owner add @shamb0t`. Then, add them to the maintainers field in this PR, or ping me to do it. Thank you.